セグ木の高速化，yakusu_enumの型変更

+++ storage/cpp/yakusu_enum.cpp
@@ -1,9 +1,9 @@
 // 約数列挙
 
-vector<int> enum_div(int n){
-	vector<int> ret;
-	for(int i = 1 ; i*i <= n ; ++i){
-		if(n%i == 0){
+vector<ll> enum_div(ll n) {
+	vector<ll> ret;
+	for(ll i = 1 ; i*i <= n ; ++i){
+		if(n%i == 0) {
 			ret.push_back(i);
 			if(i != 1 && i*i != n){
 				ret.push_back(n/i);
